refactor(view): 게임준비와 게임진행 view를 분리
