+++ src/main/java/janggi/Application.java
@@ -6,24 +6,24 @@ import janggi.dto.SetInfoDto;
 import janggi.service.GameSetDBService;
 import janggi.service.JanggiDBService;
 import janggi.util.ConnectionUtil;
-import janggi.view.InputView;
-import janggi.view.OutputView;
+import janggi.view.GameRunningView;
+import janggi.view.GameSettingView;
 import java.sql.Connection;
 
 public class Application {
     public static void main(String[] args) {
-        InputView inputView = new InputView();
-        OutputView outputView = new OutputView();
+        GameSettingView gameSettingView = new GameSettingView();
+        GameRunningView gameRunningView = new GameRunningView();
 
         Connection connection = ConnectionUtil.getConnection();
 
         GameSetDBService gameSetDBService = new GameSetDBService(connection);
-        GameSetController gameSetController = new GameSetController(inputView, outputView, gameSetDBService);
+        GameSetController gameSetController = new GameSetController(gameSettingView, gameSetDBService);
         SetInfoDto setInfoDto = gameSetController.setJanggiGame();
 
         JanggiDBService janggiDBService = new JanggiDBService(connection, setInfoDto.boardId(), setInfoDto.roomId());
 
-        JanggiController controller = new JanggiController(inputView, outputView, janggiDBService, setInfoDto.janggiGame());
+        JanggiController controller = new JanggiController(gameRunningView, janggiDBService, setInfoDto.janggiGame());
         controller.run();
     }
